# APISpec.yaml
openapi: 3.0.3
info:
  title: FTP Users
  version: 1.0.0.0
  description: APIs to Manage FTP User Accounts
servers:
  - url: 'https://api.server.test'
    description: Test API Server
  - url: 'https://api.server.prod'
    description: Production API Server
paths:
  /login:
    post:
      summary: Verify FTP User
      operationId: post-login
      description: Verify that the username/password combination is valid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples:
                ex-success:
                  value:
                    id: 132
                    status: 1
                    username: test-user
                    password: test-password
                    description: test-description
                    permissions:
                      "/":
                        - list
                        - download
                    virtual_folders:
                      - name: '1234'
                        filesystem:
                          provider: 3
                          azblobconfig:
                            container: test-container
                            account_name: test-account
                            key_prefix: 1234/
                            account_key:
                              status: Plain
                              payload: lT563yhnaanK1t7x7pZw22==
                              additional_data: folder_1234
                        virtual_path: "/1234"
                      - name: '5678'
                        filesystem:
                          provider: 3
                          azblobconfig:
                            container: test-container
                            account_name: test-account
                            key_prefix: 5678/
                            account_key:
                              status: Plain
                              payload: lT563yhnaanK1t7x7pZw22==
                              additional_data: folder_5678
                        virtual_path: "/5678"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-unauthorized:
                  value:
                    status: 401
                    location: source-file.go
                    message: Unauthorized (Failed Authentication)
                    error:
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-error:
                  value:
                    status: 500
                    location: source-file.go
                    message: Internal Server Error
                    error: Additional Error Messages
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
  /ftpusers:
    get:
      summary: Retrieve FTP Users
      operationId: get-ftpusers
      description: Returns an object with a key containing an array of all FTP User entries defined by the optional page, page_size and q(search) query parameters. The total item and total pages are also returned.
      parameters:
        - schema:
            type: integer
          in: query
          name: page
          description: (Optional) The index of the page within the ftp_users collection to return
        - schema:
            type: integer
          in: query
          name: page_size
          description: (Optional) The desired size of the result set to return
        - schema:
            type: string
          in: query
          name: q
          description: (Optional) A text string used to search the username and description fields producing a filtered result set
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FTPUsers'
              examples:
                ex-success:
                  value:
                    ftpusers:
                      - id: 1
                        username: user1
                        description: description1
                      - id: 2
                        username: user2
                        description: description2
                    total_items: 243
                    total_pages: 31
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-unauthorized:
                  value:
                    status: 401
                    location: source-file.go
                    message: Unauthorized (Failed Authentication)
                    error:
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-error:
                  value:
                    status: 500
                    location: source-file.go
                    message: Internal Server Error
                    error: Additional Error Messages
      security:
        - ApiKeyAuth: []
    post:
      summary: Create a New FTP User
      operationId: post-ftpusers
      description: Adds a new FTP User entry
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FTPUserNoPassword'
              examples:
                ex-created:
                  value:
                    id: 13
                    username: user13
                    description: description13
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-bad-request:
                  value:
                    status: 400
                    location: source-file.go
                    message: Bad Request
                    error: Username, Description and Passowrd are all required
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-unauthorized:
                  value:
                    status: 401
                    location: source-file.go
                    message: Unauthorized (Failed Authentication)
                    error:
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-bad-request:
                  value:
                    status: 409
                    location: source-file.go
                    message: Conflict
                    error: An FTP Account for [username] already exists
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-error:
                  value:
                    status: 500
                    location: source-file.go
                    message: Internal Server Error
                    error: Additional Error Messages
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FTPUser'
  '/ftpusers/{id}':
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: The id of the FTP User entry
    get:
      summary: Retrieve FTP User
      operationId: get-ftpusers-id
      description: Retrieve the FTP User entry related to {id}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FTPUserNoPassword'
              examples:
                ex-success:
                  value:
                    id: 12
                    username: user12
                    description: description 12
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-bad-request:
                  value:
                    status: 400
                    location: source-file.go
                    message: Bad Request
                    error: Invalid FTP User ID
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-unauthorized:
                  value:
                    status: 401
                    location: source-file.go
                    message: Unauthorized (Failed Authentication)
                    error:
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-not-found:
                  value:
                    status: 404
                    location: source-file.go
                    message: No matching user found
                    error:
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-error:
                  value:
                    status: 500
                    location: source-file.go
                    message: Internal Server Error
                    error: Additional Error Messages
      security:
        - ApiKeyAuth: []
    put:
      summary: Update FTP User
      operationId: put-ftpusers-id
      description: 'Updates the FTP User entry related to {id}'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FTPUserNoPassword'
              examples:
                ex-success:
                  value:
                    id: 14
                    username: user14
                    description: description 14
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-bad-request:
                  value:
                    status: 400
                    location: source-file.go
                    message: Username and Description are both required
                    error:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-unauthorized:
                  value:
                    status: 401
                    location: source-file.go
                    message: Unauthorized (Failed Authentication)
                    error:
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-not-found:
                  value:
                    status: 404
                    location: source-file.go
                    message: No matching FTP Account found
                    error:
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-error:
                  value:
                    status: 500
                    location: source-file.go
                    message: Internal Server Error
                    error: Additional Error Messages
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FTPUserUpdate'
    delete:
      summary: Delete FTP User
      operationId: delete-ftpusers-id
      description: 'Delete the FTP User entry related to {id} '
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-bad-request-invalid:
                  value:
                    status: 400
                    location: source-file.go
                    message: Invalid FTP User ID
                    error:
                ex-bad-request-conversion:
                  value:
                    status: 400
                    location: source-file.go
                    message: Cannot convert %s to an integer
                    error: Additional Error Messages
                
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-unauthorized:
                  value:
                    status: 401
                    location: source-file.go
                    message: Unauthorized (Failed Authentication)
                    error:
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-not-found:
                  value:
                    status: 404
                    location: source-file.go
                    message: No matching FTP Account found
                    error:
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-error:
                  value:
                    status: 500
                    location: source-file.go
                    message: Internal Server Error
                    error: Additional Error Messages
      security:
        - ApiKeyAuth: []
    patch:
      summary: Change Password for FTP User
      operationId: patch-ftpusers-id
      description: Change the password for the FTP User related to {id}
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-bad-request-invalid:
                  value:
                    status: 400
                    location: source-file.go
                    message: Invalid FTP User ID
                    error:
                ex-bad-request-conversion:
                  value:
                    status: 400
                    location: source-file.go
                    message: Cannot convert %s to an integer
                    error: Additional Error Messages
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-unauthorized:
                  value:
                    status: 401
                    location: source-file.go
                    message: Unauthorized (Failed Authentication)
                    error:
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-not-found:
                  value:
                    status: 404
                    location: source-file.go
                    message: No matching FTP Account found
                    error:
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ex-error:
                  value:
                    status: 500
                    location: source-file.go
                    message: Internal Server Error
                    error: Additional Error Messages
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FTPUserPassword'
...
components:
  schemas:
    Error:
      title: Error
      type: object
      properties:
        status:
          type: integer
          description: The HTTP Status returned with the header in the response
        location:
          type: string
          description: A reference to the location in the code where the error occurred
        message:
          type: string
          description: A custom error message generated by the application
        error:
          type: string
          description: An error message generated by the runtime environment
      required:
        - status
        - location
    Credentials:
      title: Credentials
      type: object
      description: A set of credentials to verify
      properties:
        username:
          type: string
          description: The username of the FTP User entry
        password:
          type: string
          description: The password of the FTP User entry
      required:
        - username
        - password
    FTPUser:
      title: FTPUser
      type: object
      description: An FTP User entry
      properties:
        username:
          type: string
          description: The username for the FTP User
        description:
          type: string
          description: The description for the FTP User
        password:
          type: string
          description: The password for the FTP User
      required:
        - username
        - description
        - password
    FTPUserNoPassword:
      title: FTPUserNoPassword
      type: object
      description: An FTP User entry without the password
      properties:
        id:
          type: integer
          description: The FTP User entry id
        username:
          type: string
          description: The username for the FTP User
        description:
          type: string
          description: The description for the FTP User
      required:
        - id
        - username
        - description
    FTPUsers:
      title: FTPUsers
      type: object
      properties:
        ftpusers:
          type: array
          items:
            $ref: '#/components/schemas/FTPUserNoPassword'
        total_items:
          type: integer
        total_pages:
          type: integer
      description: A collection of FTPUser records without the password.  The count of ftpuser items is provided as well as the total pages of items.
    FTPUserUpdate:
      title: FTPUserUpdate
      type: object
      description: An FTP User update entry
      properties:
        username:
          type: string
          description: The username for the FTP User
        description:
          type: string
          description: The description for the FTP User
      required:
        - username
        - description
    FTPUserPassword:
      title: Password
      type: object
      description: A password used in an update
      properties:
        password:
          type: string
          description: The password of the FTP User entry
      required:
        - password
...
    Permission:
      type: string
      enum:
        - '*'
        - list
        - download
        - upload
        - overwrite
        - delete
        - delete_files
        - delete_dirs
        - rename
        - rename_files
        - rename_dirs
        - create_dirs
        - create_symlinks
        - chmod
        - chown
        - chtimes
      description: |
        Permissions:
          * `*` - all permissions are granted
          * `list` - list items is allowed
          * `download` - download files is allowed
          * `upload` - upload files is allowed
          * `overwrite` - overwrite an existing file, while uploading, is allowed. upload permission is required to allow file overwrite
          * `delete` - delete files or directories is allowed
          * `delete_files` - delete files is allowed
          * `delete_dirs` - delete directories is allowed
          * `rename` - rename files or directories is allowed
          * `rename_files` - rename files is allowed
          * `rename_dirs` - rename directories is allowed
          * `create_dirs` - create directories is allowed
          * `create_symlinks` - create links is allowed
          * `chmod` changing file or directory permissions is allowed
          * `chown` changing file or directory owner and group is allowed
          * `chtimes` changing file or directory access and modification time is allowed
    DirPermissions:
      type: object
      additionalProperties:
        type: array
        items:
          $ref: '#/components/schemas/Permission'
        minItems: 1
      minProperties: 1
      description: 'hash map with directory as key and an array of permissions as value. Directories must be absolute paths, permissions for root directory ("/") are required'
    FsProviders:
      type: integer
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
      description: |
        Filesystem providers:
          * `0` - Local filesystem
          * `1` - S3 Compatible Object Storage
          * `2` - Google Cloud Storage
          * `3` - Azure Blob Storage
          * `4` - Local filesystem encrypted
          * `5` - SFTP
...
    CryptFsConfig:
      type: object
      properties:
        passphrase:
          $ref: '#/components/schemas/Secret'
      description: Crypt filesystem configuration details
...
    FilesystemConfig:
      type: object
      properties:
        provider:
          $ref: '#/components/schemas/FsProviders'
        s3config:
          $ref: '#/components/schemas/S3Config'
        gcsconfig:
          $ref: '#/components/schemas/GCSConfig'
        azblobconfig:
          $ref: '#/components/schemas/AzureBlobFsConfig'
        cryptconfig:
          $ref: '#/components/schemas/CryptFsConfig'
        sftpconfig:
          $ref: '#/components/schemas/SFTPFsConfig'
      description: Storage filesystem details
    BaseVirtualFolder:
      type: object
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
        name:
          type: string
          description: unique name for this virtual folder
        mapped_path:
          type: string
          description: absolute filesystem path to use as virtual folder
        description:
          type: string
          description: optional description
        used_quota_size:
          type: integer
          format: int64
        used_quota_files:
          type: integer
          format: int32
        last_quota_update:
          type: integer
          format: int64
          description: Last quota update as unix timestamp in milliseconds
        users:
          type: array
          items:
            type: string
          description: list of usernames associated with this virtual folder
        filesystem:
          $ref: '#/components/schemas/FilesystemConfig'
      description: 'Defines the filesystem for the virtual folder and the used quota limits. The same folder can be shared among multiple users and each user can have different quota limits or a different virtual path.'
    VirtualFolder:
      allOf:
        - $ref: '#/components/schemas/BaseVirtualFolder'
        - type: object
          properties:
            virtual_path:
              type: string
            quota_size:
              type: integer
              format: int64
              description: 'Quota as size in bytes. 0 menas unlimited, -1 means included in user quota. Please note that quota is updated if files are added/removed via SFTPGo otherwise a quota scan or a manual quota update is needed'
            quota_files:
              type: integer
              format: int32
              description: 'Quota as number of files. 0 menas unlimited, , -1 means included in user quota. Please note that quota is updated if files are added/removed via SFTPGo otherwise a quota scan or a manual quota update is needed'
          required:
            - virtual_path
      description: 'A virtual folder is a mapping between a SFTPGo virtual path and a filesystem path outside the user home directory. The specified paths must be absolute and the virtual path cannot be "/", it must be a sub directory. The parent directory for the specified virtual path must exist. SFTPGo will try to automatically create any missing parent directory for the configured virtual folders at user login.'
    User:
      type: object
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
        status:
          type: integer
          enum:
            - 0
            - 1
          description: |
            status:
              * `0` user is disabled, login is not allowed
              * `1` user is enabled
        username:
          type: string
          description: username is unique
        email:
          type: string
          format: email
        description:
          type: string
          description: 'optional description, for example the user full name'
        expiration_date:
          type: integer
          format: int64
          description: expiration date as unix timestamp in milliseconds. An expired account cannot login. 0 means no expiration
        password:
          type: string
          format: password
          description: password or public key/SSH user certificate are mandatory. If the password has no known hashing algo prefix it will be stored, by default, using bcrypt, argon2id is supported too. You can send a password hashed as bcrypt ($2a$ prefix), argon2id, pbkdf2 or unix crypt and it will be stored as is. For security reasons this field is omitted when you search/get users
        public_keys:
          type: array
          items:
            type: string
            example: ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBEUWwDwEWhTbF0MqAsp/oXK1HR2cElhM8oo1uVmL3ZeDKDiTm4ljMr92wfTgIGDqIoxmVqgYIkAOAhuykAVWBzc= user@host
          description: Public keys in OpenSSH format. A password or at least one public key/SSH user certificate are mandatory.
        home_dir:
          type: string
          description: path to the user home directory. The user cannot upload or download files outside this directory. SFTPGo tries to automatically create this folder if missing. Must be an absolute path
        virtual_folders:
          type: array
          items:
            $ref: '#/components/schemas/VirtualFolder'
          description: mapping between virtual SFTPGo paths and filesystem paths outside the user home directory. Supported for local filesystem only. If one or more of the specified folders are not inside the dataprovider they will be automatically created. You have to create the folder on the filesystem yourself
        uid:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
          description: 'if you run SFTPGo as root user, the created files and directories will be assigned to this uid. 0 means no change, the owner will be the user that runs SFTPGo. Ignored on windows'
        gid:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
          description: 'if you run SFTPGo as root user, the created files and directories will be assigned to this gid. 0 means no change, the group will be the one of the user that runs SFTPGo. Ignored on windows'
        max_sessions:
          type: integer
          format: int32
          description: Limit the sessions that a user can open. 0 means unlimited
        quota_size:
          type: integer
          format: int64
          description: Quota as size in bytes. 0 menas unlimited. Please note that quota is updated if files are added/removed via SFTPGo otherwise a quota scan or a manual quota update is needed
        quota_files:
          type: integer
          format: int32
          description: Quota as number of files. 0 menas unlimited. Please note that quota is updated if files are added/removed via SFTPGo otherwise a quota scan or a manual quota update is needed
        permissions:
          type: object
          items:
            $ref: '#/components/schemas/DirPermissions'
          minItems: 1
          example:
            /:
              - '*'
            /somedir:
              - list
              - download
        used_quota_size:
          type: integer
          format: int64
        used_quota_files:
          type: integer
          format: int32
        last_quota_update:
          type: integer
          format: int64
          description: Last quota update as unix timestamp in milliseconds
        upload_bandwidth:
          type: integer
          description: 'Maximum upload bandwidth as KB/s, 0 means unlimited'
        download_bandwidth:
          type: integer
          description: 'Maximum download bandwidth as KB/s, 0 means unlimited'
        upload_data_transfer:
          type: integer
          description: 'Maximum data transfer allowed for uploads as MB. 0 means no limit'
        download_data_transfer:
          type: integer
          description: 'Maximum data transfer allowed for downloads as MB. 0 means no limit'
        total_data_transfer:
          type: integer
          description: 'Maximum total data transfer as MB. 0 means unlimited. You can set a total data transfer instead of the individual values for uploads and downloads'
        used_upload_data_transfer:
          type: integer
          description: 'Uploaded size, as bytes, since the last reset'
        used_download_data_transfer:
          type: integer
          description: 'Downloaded size, as bytes, since the last reset'
        created_at:
          type: integer
          format: int64
          description: 'creation time as unix timestamp in milliseconds. It will be 0 for users created before v2.2.0'
        updated_at:
          type: integer
          format: int64
          description: last update time as unix timestamp in milliseconds
        last_login:
          type: integer
          format: int64
          description: Last user login as unix timestamp in milliseconds. It is saved at most once every 10 minutes
        filters:
          $ref: '#/components/schemas/UserFilters'
        filesystem:
          $ref: '#/components/schemas/FilesystemConfig'
        additional_info:
          type: string
          description: Free form text field for external systems
        oidc_custom_fields:
          type: object
          additionalProperties: true
          description: 'This field is passed to the pre-login hook if custom OIDC token fields have been configured. Field values can be of any type (this is a free form object) and depend on the type of the configured OIDC token fields'
  securitySchemes:
    ApiKeyAuth:
      name: X-API-KEY
      type: apiKey
      in: header